requirements.txt
pytest==7.4.3
playwright==1.58.0
allure-pytest==2.13.2
pytest-html==4.1.1
pytest-xdist==3.5.0
python-dotenv==1.0.1